feat(password-update): Pass username and old password for temp. auth.
The REST resources support temporary authentication for validation and
update requests if username and password is given in request body.
We use this feature here if `username` and `oldPassword` is given via
input parameters. If they're not set, we expect the current user to be
authenticated and to be permitted to execute the REST calls (so no
extra authentication needed).