fix(entity-browser): Remove custom cell renderer
The renderer doesn't work anymore, because the signature changed (array of
fields is given as first argument, instead of a single field).
We could fix the renderer, but we can also remove it, since the only difference
to the default renderer is, that this custom renderer returns a `_` instead of
nothing. We don't really need that.