refactor(tocco-ui): call the onSearch function also if the value is empty
If the value of the input had more characters than defined with the
minInputLength property and the user clears the input afterwards
the onSearch function was not called which would make sense.
So if for ex. the `SearchBox` component is used to filter a list and
a user types some input the `onSearch` function will be called.
If the user afterwards clears the input again he expects to have the
list reset.