feat(entity-browser): add handling to handle a different forms
If no form base is set, the forms will be requested by the entity name. If a base form is set, this base will be taken to request the search and list form.
For ex.: if the form base is set ti `My_form` the EntityBrowser will request the two forms `My_form_search` and `My_form_list`.