fix(tocco-ui): fix a date formatter time zone issue
Before the time zone was set to UTC fixed. This caused an issue because dates in js alsways
have a time and a time zone. Different browser handle the time zone differently. Some
use the UTC time zone and some use the local time zone (e.g. GMT +2)

Now the time zone is not set to UTC by default anymore. If a date object is passed to the
date formatter, it just will be passed. In this case the time zone handling does not matter.

If a string or a number is passed, the `Date.parse` function is used to parse the date and
the time zone offset will be added to the created date object.