feat(entity-browser): Confirmation dialog if touched form is left
To show a confirmation box, only the `<Prompt>` component has to be added with
the desired message.

However, by default, the native confirmation dialog of the browser is displayed
which is not that pretty. To display a nicer toastr confirmation box, we have
to use the `getUserConfirmation` option when we create the history.

At the moment, there is an ongoing discussion about how to use a custom
component for the dialog here:
https://github.com/ReactTraining/react-router/issues/4635