chore(tocco-client): switch license to AGPLv3+ (adding the +)
Omitting the + means we can't upgrade to a new version of the
license should the copyright holder of any commit ever be any
other party than Tocco. Also, we really want to be able to
upgrade in case the license turns out to be flawed, legally
unenforcable or if legal circumstances change.

note: SPDX allows a + suffix to indicate "or any later version"