refactor(entity-detail): rewrite getDirtyFields
when loaded in nice2 context, getDirtyFields always returned an empty array.
It seems like the reduce function from lodash internally uses a function, that behaves differently when running in nice2 context.
Probably a polyfill overwrites a functionality.
It couldn't be analyzed what exactly caused this so in a simple workaround the whole method got rewritten to work without _reduce

Signed-off-by: Daniel Keller <dkeller@tocco.ch>