feat(entity-list): Refresh the current page on mount if initialized
The `initialize` saga is called when the list view is mounted.

In case it's already initialized, we should still refresh the current
page, since the displayed data might be outdated now (example: user
switches from the list view to the detail view of an entity, updates
the entity and returns back to the list -> list is now outdated and
should be refreshed).