feat(entity-list): add row selection feature
Rows now can be selected if the property `selectable` is set to `true`
on the ListView component.
The selection model provides either to select/deselect single rows or
select/deselect all visible rows at the same time.
All selected rows are persisted in the redux state (state.list.selection).

A selection also can be passed as array of numbers to the ListView component
to preselect rows or to overwrite an existing selection.