feat(entity-list): include search filter option
Search filters can now be included in search forms. If a search filter
field is included, the search request will be adjusted if a search
filter is selected. The `_filter` argument will be added to the request.
It is possible to select multiple search filters. If so, the conditions
will be combined.