chore(scheduler): use raw html in calendar event popover
the description property of a calendar event now will return html
instead of text. to correctly apply the html to the popover the
`dangerouslySetInnerHtml` property for React components is used.

the popover content has to be split because in raw html code the react
component `Conflict` will not be recognized otherwise.