fix(tocco-util): prevent scrollbar while bounce in notifier
CSS animation with scale in combination with "overflow: auto;" are problematic. Some browser-device-combinations display a scrollbar during the animation.

Safeares created by padding or margin are not predictable because scaling is a precentage of the intrinsic height.
Disabling overflow temporally is not straigth forward because it has to be declared on the parent element of the animate element.

The implemented solution is a animation which never causes an overflow because it respect the safeareas by default.