refactor(resource-scheduler): Disable selection controller in lists
It doesn't make sense to select a big number of entities in the search
panel lists to display them in the calendar. The calendar component
won't be able to display a big number of entities, so we don't have
to support the new "select all" feature. Everything will work just
like before in the resource scheduler, once the selection feature
branch has been merged into the master branch.