refactor(simple-form): Add `changeFieldValue` to form utils
Could, for instance, be used in `editableValueFactory` to change some
other field value on certain events.

Rather stupid example which defines a wrapper for the `onChange` event
handler of all components in the `editableValueFactory` to change the
value of `lastname` to "foobar" whenever a form value is changed:

```
const onChangeFn = props.onChange
props.onChange = (...args) => {
  onChangeFn(...args)
  utils.changeFieldValue('lastname', 'foobar')
}
```

A real world example would be the new location field which updates
other fields like district and country when a city/zipcode gets
selected.

The redux-form action `change` is used which takes three arguments:
- redux form name
- field name
- field value

In order to be able to call `changeFieldValue` with only the field name
and the value, we bind the redux form name as the first argument
in the `Form` component (where the redux form is created using
the `reduxForm` HOC). In the other places where `reduxForm` is used
in our code base, the `changeFieldValue` function can be added
to the form utils in the exact same way.