refactor(tocco-ui): harmonize props to disable fields
The problem is that props readonly and disabled are used for different things.
- If these props are valid for the HTML tag it is added to the DOM which is not obvious because of implementation in two separate files.
- The styles for immutable field are always identical, independently of the HTML attribute or if a third party library is used.

Hence the styles must be independent of the particular implementation and the props must be passed correctly and comprehensible to all components.

The solution is to harmonize all the names (disabled, isDisabled, immutable, readOnly) as far as possible, to use primarily a name which can not be interpreted as HTML attribute, to be explicit about declaring HTML attributes and to style without dependencies on attributes.

Caveats are that the backend use readonly which is mapped in typeEditorFactory to immutable and that third party libraries implement immutability in various ways.