feat(admin): include merge action in admin app
The merge action is included as legacy action in the admin app at the
moment. To be able to run, the `merge` React app needs to be registered
in the global `reactRegistry` which is achieved by importing the
merge app (registers itself when loaded).

The change https://git.tocco.ch/c/nice2/+/36857 actually adds the
button to the forms to trigger the action.

Refs: TOCDEV-1878