fix(admin): Take `baseRoute` into account for resource scheduler
For the deployment on master.tocco.ch, the admin input prop `baseRoute`
(which is used as `basename` for the router) is set to `/beta`.
By using `history#createHref` the URL is prefixed with the basename
which was missing before -> the entities that should be opened when
the user clicks on an event in the scheduler couldn't be displayed
on master.tocco.ch because the URL was wrong.

Refs: TOCDEV-2347