feat(entity-list): Allow parent to be changed via input
It was already possible to set the initial parent of the entity list
via the `parent` input argument.

However, when the entity-list was rerendered with a different parent,
this change was ignored.

To take these changes into account, we use the parent from the
`entityList` state which contains the updated input argument to
determine the REST endpoint and we also call the `reloadData` saga
whenever the parent is updated.

We also have to consider `null` as a defined value in input.js,
otherwise it's not possible to reset the current parent to null.

Refs: TOCDEV-2560