feat(entity-list): Add support for `searchEndpoint` table attribute
The `searchEndpoint` attribute can be used to override the `endpoint`
attribute for search requests (search requests means: a request with
search attributes *defined by the user* -> the input TQL conditions
don't count, for example).

This will, for instance, be used in the DMS, where we have to fetch
search results from an endpoint which returns all nodes regardless
of their actual location in the tree.

Refs: TOCDEV-2560