feat(tocco-ui): Set fixed width on Fontawesome icons
This is, for example, useful in the DMS:
The label column consists of an icon depending on the type and the
actual label. If we don't use the `fixedWidth` feature, the labels
in the list aren't aligned properly, because the icons in front of
them don't have the same width.

Changelog: Use same width for all Fontawesome icons
Refs: TOCDEV-2882