feat(entity-list): Add support for custom cell renderers
Custom cell renderers can be specified via the `client-renderer`
column attribute in the XML definition
(also see https://git.tocco.ch/c/nice2/+/39168).

The actual renderer can be passed as `cellRenderers` input to the
entity list as a map with the name of the renderer as key and the
rendering function as value. The rendering function receives the
row data (= entity), the column definition and the default rendering
function. The given rendering function can be used to render certain
children of the column using the default mechanism.

This cell renderers are, for example, useful in the DMS:
The label column consists of an icon depending on the type and the
actual label. The icon will be inserted by a cell renderer in the
label column.

Changelog: Introduce custom cell renderers
Refs: TOCDEV-2882