fix(entity-detail): Fix writing to fields in related entities
Cause of the bug:
Fields in related entities are contained in the flat entity
representation in the format '{relation name}.{field name}',
but the format in the list of the dirty fields is
'{relation name}--{field name}'.
Therefore, those fields are never considered dirty and thus those
fields are never added to the request.

Fix:
We have to convert the dirty field names back to
'{relation name}.{field name}' before calling `api.toEntity`.

Example (there aren't many of them):
One example where we want to write to a field in a related entity is
`relContent.description` (among other fields on `relContent`) in
`relResource_detail`.

Refs: TOCDEV-3118