fix(app-extension): Fix legacy actions that call `refreshListPanel()`
Until now, the action context didn't have a panel (2nd constructor
parameter of the action context). This led to a null pointer when
the legacy action tried to call
`this.getActionContext().getPanel().getStore().reload()` in
`refreshListPanel()`.
Now we're providing a fake panel with a fake store to avoid the null
pointer exception and to reload the list.

Refs: TOCDEV-3100