feat(two-factor-connector): Add external event `onResize`
This event is fired whenever the size of the screen (resp. its height)
changes substantially (similar to the `resize` event of the login
and password update app).

This event, for instance, is useful if the app is rendered in a popup
window which should be centered whenever the height of the app changes.

Refs: TOCDEV-3048