feat(docs-browser): Add input param `rootNodes` to define root items
If `rootNodes` are defined, they will be displayed in the root list
instead of the domains.
If the `rootNodes` array contains exactly *one* item, it will be used
as the "root parent" and thus the root list contains the children of
the single root node.

Refs: TOCDEV-3368
Changelog: New input param `rootNodes` to define root items