chore: scripts do not push temporary release tags
if multiple auto cherrypicking ci jobs (on different release branches) are executed "git push --tags" fails because an auto-merge tag (of another nice-release) was moved. As the pushed tags with the prefix "_" are deleted by the tagging ci job now the tags are instant locally deleted after the creation.

Cherry-pick: Up