fix(devcon): Prevent complete reload of DevCon when changing page
"Seamless" navigation enabled by using the `RouterLink` component
instead of the `Link` component (the latter renders a simple `<a>`
Link instead of a RouterLink).

This means, that the Redux store is kept and that we can navigate
around in the DevCon without losing application log or SQL log entries
etc.

Refs: TOCDEV-3951