fix(devcon): Fix generating SQL and changesets in prod env
Requesting the SQL and changesets with the `selection` (which
is a Set) for some reason works in local dev environment (package
`devcon` started locally in tocco-client repository), but gets
serialized to `{}` when the app is included in the Nice repository,
which leads to an error because a List<String> is expected on the
server side.

We have to convert it to an array to fix the serialization.

Refs: TOCDEV-3951
Changelog: Fix generating SQL and changesets in prod env