fix(entity-browser): fix opening entity-browser the second time
The entity-browser was broken when it was
opened the second time by e.g. navigating away and coming back to it.
The problem was that it got re-rendered with the latest persisted state as
soon as something changed. Therefore it triggered the refresh data action
instead of handling it as a normal initialisation.

Changelog: fix opening entity-browser the second time
Refs: TOCDEV-4383