fix(entity-list): set store initially vs. re-init on store updates
Fixes an error when a prop changes during initialization phase. In that phase the
`store` prop has been changed from `undefined` to defined as well which triggers a re-initialization of the entity-list.
In this case the entity-list should not handle this update as store change and should
ignore the passed store. Therefore we distinguish between setting the store initially for
recovering previous state (e.g. navigating from detail back to list) or
if we really want to re-initialize the entity-list when we explicitly
set another store to the app (e.g. in the docs-browser for navigating between levels).

Changelog: set store initially vs. re-init on store updates
Refs: TOCDEV-4465