fix(entity-list): only reload on explicit store updates
Fixes an error when a prop changes during initialization phase. In that phase the
store prop has been changed from `undefined` to defined as well which triggers a re-initialization of the entity-list.
In this case the entity-list should not handle this update as store change and should
ignore the passed store. Therefore we distinguish between setting the store after being initalized with another store,
or when setting the store the first time because of an update from another prop.

Changelog: only reload on explicit store updates
Refs: TOCDEV-4465