fix(entity-list): reset pagination on reload first page
Whenever the query for fetching the data changes the data
get fetched for the first page explicitly. However the pagination
(current page) has not been reset. Therefore the pagnination
showed wrong page information and refreshing the page was
not possible in some cases.

Changelog: reset pagination on reload first page
Refs: TOCDEV-4467