fix(app-extension): Set correct value after creation in remote field
After a new record was created directly in a remote field via the "+" button,
the new record was set as value.

This was the correct behavior for single-remote-fields. However, it led to
the error that is described in TOCDEV-4546 in multi-remote-fields.

In multi-remote-fields, we have to set an *array* value instead of a single object.
Furthermore, the new object should be *appended* to the entries that are already
set in the field (they shouldn't be overridden completely).

Refs: TOCDEV-4546
Changelog: Fix creating related entities in multi-remote-fields