fix(app-extension): remove queryOrderBy from legacy action selection
a legacy action didn't successfully finished if no selection was made and the list was sorted by a column which belongs to a relation (for example if the reservation list was sorted by the relReservation_status). for legacy actions just the same sorting string was used as for the entities 2.0 rest endpoint. however the endpoint supports as sorting string "relReservation_status asc" while the legacy client does only support fields (e.g. "relReservation_status.label asc"). if in the legacy client the list is sorted and an action called the sorting is not passed to the backend (only if query tab is used).

Refs: TOCDEV-4085
Changelog: remove queryOrderBy from legacy action selection