fix(admin): set focus to search input on menu change
As soon as the menu tab changes the focus should
be set to the serach input. No matter if it has been
changed by clicking on the menu icon or by using
the shortcut. It was only working by clicking on the icon.

Changelog: set focus to search input on menu change
Refs: TOCDEV-4551