feat: add component name as CSS classname
Makes it possible to style elements by CSS from the outside
when our apps are embedded as a widget in external websites.

Caution:

The class names look like this:
{Component name}-Tocco__sc-{Random suffix}

Because of the random suffix, the elements can't be selected using
the regular syntax (e.g. .foobar), but the elements should be selected
via an attribute selector (e.g. [class*=StyledTable-Tocco]).

Caution 2:

This way of styling a widget is very unstable, as class names and DOM
structure might evolve over time. CSS rules should only be defined
if really necessary and if it's not possible to style something via the
theme object.

Refs: TOCDEV-388, TOCDEV-4758
Changelog: add component name as CSS classname