feat(tocco-util): add `localeSelector` to `intl` util
This selector should be used to get the current locale of the app.

Do not write your own selector that reads the locale from `input`
instead of `intl` as it might be outdated (the locale of the app
could have been changed using `changeLocale`).

Also keep in mind: if you map the state to your component's
props in a Redux container, you should also use `state.intl.locale`
instead of `state.input.locale`.

Refs: TOCDEV-4817