feat(entity-list): add prop to use separate list form for search mode
The list form definition that should be used as soon as the user
starts to search can be defined by the input prop `searchListFormName`.
If not defined, the regular list form is used.

Needed for the DMS for the following reasons:
- The first level shows the domains and there should also be a column
  for the type of the domain. Once the user starts to search, there
  not only domains displayed, but also folders and resources;
  therefore, this column should not be visible then anymore.
- Once the user starts to search, a "flat" list is displayed - there's no
  parent node anymore. Therefore, the "new" actions don't make sense
  here and we don't want to display them.

Refs: TOCDEV-5168, TOCDEV-5167
Changelog: new prop `searchListFormName` to use separate list form for search mode