fix(admin): fix navigating to entity document
Since TOCDEV-5060 and commit c49a167 navigating to a entity document
via the hash URL didn't work anymore, as the hash URL was ignored since
then ("routerless" docs-browser instance is used, as routers cannot be
nested anymore).

To fix navigating to the document, the hash URL is passed as
`initialLocation` to the <DocsBrowser> now.

Refs: TOCDEV-5343, TOCDEV-5060
Changelog: fixed navigating to entity document