fix(admin): do not re-initialize admin on session heartbeat
The session heartbeat is setup after login but will not be setup again
on a page refresh with having still a valid session.
The nice2 session timeout is set to 30 minutes but the
frontend assumed to have a session timeout of 10s.
After 22min the session heartbeat is executed, but only
when tab hasn't been refreshed after login.
The admin got completely umounted and mounted on every heartbeat.
This caused unsaved changes to be lossed completely.
To not show "no-roles" error messages right after login but
before the session has beend fetched, the `adminAllowed` variable
has to be reset after login.
This prevents to render the admin with a value of `false` and showing
the error message for a short time.

Refs: TOCDEV-5100
Changlog: do not re-initalize admin on session heartbeat
Cherry-pick: Up