feat(app-extensions): add condition to remote and select field
on remote and select fields a dynamic condition can be set. this allows to add conditions based on the selection (e.g. entity model) in simple actions (important forms are cached so conditions are not part of the form model and are set later)

Refs: TOCDEV-5584
Changelog: add condition to remote and select field