fix(tocco-ui): HTML editor: fix saving the form in source editing mode
The 'SourceEditing' plugin internally updates the editor sources
on `editor.getData()`. This leads to an infinite loop of
`getData()` calls and `onChange` events (as this `onChange`
event handler calls `getData()`.) Therefore, we have to make
sure that we handle the `onChange` only once if we're in the
source editing mode.

Note:
The `onChange` never actually gets fired during changes in the
source editing mode. Rather, the editor gets unmounted before the form
gets saved. During this procedure an internal destroy function is called
which updates the sources and in turn fires the `onChange` event.

Note 2:
In general, it seems quite fragile and hacky. However, the SourceEditing
plugin exposes only the `isSourceEditingMode` and there are no useful
events or something like that. So, this solution is the best option I've
come up with without accessing private fields of the plugin or writing
our own source editing plugin.

Refs: TOCDEV-5618
Changelog: HTML editor: fix saving the form in source editing mode
Cherry-pick: Up