feat(entity-browser): add action callbacks to navigate back
The standard action events `onSuccess`, `onError` and `onCancel`
navigate back to the list or detail view (depending on where
the action was opened from). This is the same behavior as in the
admin app. The entity-browser app is used for many widgets
(e.g. the input edit widget where we're going to add a "back"
button in the referenced Jira task).

Refs: TOCDEV-5871
Changelog: added action callbacks to navigate back