fix(tocco-util): fix performance of fulltext search
just using one fulltext keyword instead of two has a better performance. in the old client the search term "term1 term2 term3" is converted to the TQL "term1 term2 (term3 OR term3*)" which is the equivalent of the TQL  "(term1 term2 term3) OR (term1 term2 term3*)" now implemented in the new client.

Refs: TOCDEV-6080
Changelog: fix performance of fulltext search