feat(app-extensions): allow 400 status code in async validation
if a field is focused and the invalid session modal is shown the focus is changed to the modal. This triggers the async validation. The server returns a 400 status code as the user is no longer logged in. This error should be ignored

Changelog: allow 400 status code in async validation
Refs: TOCDEV-5313