fix(scheduler): fix null as end date
While the start date is inclusive, the end date is exclusive (https://fullcalendar.io/docs/event-parsing). If the same timestamp is passed for start and end of the event, the eventInfo.event.end of the eventContent method is null. Use the start date as fallback if the end date is null

Changelog: fix null as end date
Refs: TOCDEV-6246
Cherry-pick: Up